+++ Classes/AbstractDataHandlerListener.php
@@ -24,8 +24,7 @@ namespace ApacheSolrForTypo3\Solr;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
-use ApacheSolrForTypo3\Solr\System\Configuration\ConfigurationManager;
-use ApacheSolrForTypo3\Solr\System\Configuration\TypoScriptConfiguration;
+use ApacheSolrForTypo3\Solr\Domain\Index\Queue\RecordMonitor\Helper\ConfigurationAwareRecordService;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
 use TYPO3\CMS\Core\Database\QueryGenerator;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -44,13 +43,13 @@ abstract class AbstractDataHandlerListener
     /**
      * Reference to the configuration manager
      *
-     * @var \ApacheSolrForTypo3\Solr\System\Configuration\ConfigurationManager
+     * @var \ApacheSolrForTypo3\Solr\Domain\Index\Queue\RecordMonitor\Helper\ConfigurationAwareRecordService
      */
-    protected $configurationManager;
+    protected $configurationAwareRecordService;
 
     public function __construct()
     {
-        $this->configurationManager = GeneralUtility::makeInstance(ConfigurationManager::class);
+        $this->configurationAwareRecordService = GeneralUtility::makeInstance(ConfigurationAwareRecordService::class);
     }
 
     /**
@@ -113,7 +112,7 @@ abstract class AbstractDataHandlerListener
         // If RecursiveUpdateTriggerConfiguration is false => check if changeFields are part of recursiveUpdateFields
         if ($isRecursiveUpdateRequired === false) {
             $solrConfiguration = Util::getSolrConfigurationFromPageId($pageId);
-            $indexQueueConfigurationName = $this->configurationManager->getIndexingConfigurationName('pages', $pageId, $solrConfiguration);
+            $indexQueueConfigurationName = $this->configurationAwareRecordService->getIndexingConfigurationName('pages', $pageId, $solrConfiguration);
             $updateFields = $solrConfiguration->getIndexQueueConfigurationRecursiveUpdateFields($indexQueueConfigurationName);
 
             // Check if no additional fields have been defined and then skip recursive update
@@ -181,32 +180,6 @@ abstract class AbstractDataHandlerListener
 
         $diff = array_diff_assoc($triggerConfiguration['changeSet'], $changedFields);
         return empty($diff);
-    }
-
-    /**
-     * Retrieves a record, taking into account the additionalWhereClauses of the
-     * Indexing Queue configurations.
-     *
-     * @param string $recordTable Table to read from
-     * @param int $recordUid Id of the record
-     * @param TypoScriptConfiguration $solrConfiguration
-     * @return array Record if found, otherwise empty array
-     */
-    protected function getRecord($recordTable, $recordUid, TypoScriptConfiguration $solrConfiguration)
-    {
-        $record = [];
-        $indexingConfigurations = $solrConfiguration->getEnabledIndexQueueConfigurationNames();
-
-        foreach ($indexingConfigurations as $indexingConfigurationName) {
-            $record = $this->configurationManager->getRecordIfIndexConfigurationIsValid($recordTable, $recordUid,
-                $indexingConfigurationName, $solrConfiguration);
-            if (!empty($record)) {
-                // if we found a record which matches the conditions, we can continue
-                break;
-            }
-        }
-
-        return $record;
     }
 
     /**
