+++ Classes/System/Configuration/ConfigurationManager.php
@@ -107,112 +107,4 @@ class ConfigurationManager implements SingletonInterface
         return GeneralUtility::makeInstance(TypoScriptConfiguration::class, $configurationArray, $contextPageId);
     }
 
-    /**
-     * Retrieves the name of the Index Queue Configuration for a record.
-     *
-     * @param string $recordTable Table to read from
-     * @param int $recordUid Id of the record
-     * @param TypoScriptConfiguration $solrConfiguration
-     * @return string Name of indexing configuration
-     */
-    public function getIndexingConfigurationName(
-        $recordTable,
-        $recordUid,
-        TypoScriptConfiguration $solrConfiguration
-    ) {
-        $name = $recordTable;
-        $indexingConfigurations = $solrConfiguration->getEnabledIndexQueueConfigurationNames();
-        foreach ($indexingConfigurations as $indexingConfigurationName) {
-            if (!$solrConfiguration->getIndexQueueConfigurationIsEnabled($indexingConfigurationName)) {
-                // ignore disabled indexing configurations
-                continue;
-            }
-
-            $record = $this->getRecordIfIndexConfigurationIsValid($recordTable, $recordUid,
-                $indexingConfigurationName, $solrConfiguration);
-            if (!empty($record)) {
-                $name = $indexingConfigurationName;
-                // FIXME currently returns after the first configuration match
-                break;
-            }
-        }
-
-        return $name;
-    }
-
-    /**
-     * This method return the record array if the table is valid for this indexingConfiguration.
-     * Otherwise an empty array will be returned.
-     *
-     * @param string $recordTable
-     * @param integer $recordUid
-     * @param string $indexingConfigurationName
-     * @param TypoScriptConfiguration $solrConfiguration
-     * @return array
-     */
-    public function getRecordIfIndexConfigurationIsValid(
-        $recordTable,
-        $recordUid,
-        $indexingConfigurationName,
-        TypoScriptConfiguration $solrConfiguration
-    ) {
-        if (!$this->isValidTableForIndexConfigurationName($recordTable, $indexingConfigurationName,
-            $solrConfiguration)
-        ) {
-            return [];
-        }
-
-        $recordWhereClause = $solrConfiguration->getIndexQueueAdditionalWhereClauseByConfigurationName($indexingConfigurationName);
-
-        if ($recordTable === 'pages_language_overlay') {
-            return $this->getPageOverlayRecordIfParentIsAccessible($recordUid, $recordWhereClause);
-        }
-
-        return (array)BackendUtility::getRecord($recordTable, $recordUid, '*', $recordWhereClause);
-    }
-
-    /**
-     * This method is used to check if a table is an allowed table for an index configuration.
-     *
-     * @param string $recordTable
-     * @param string $indexingConfigurationName
-     * @param TypoScriptConfiguration $solrConfiguration
-     * @return boolean
-     */
-    protected function isValidTableForIndexConfigurationName(
-        $recordTable,
-        $indexingConfigurationName,
-        TypoScriptConfiguration $solrConfiguration
-    ) {
-        $tableToIndex = $solrConfiguration->getIndexQueueTableNameOrFallbackToConfigurationName($indexingConfigurationName);
-
-        $isMatchingTable = ($tableToIndex === $recordTable);
-        $isPagesPassedAndOverlayRequested = $tableToIndex === 'pages' && $recordTable === 'pages_language_overlay';
-
-        if ($isMatchingTable || $isPagesPassedAndOverlayRequested) {
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * This method retrieves the pages_language_overlay record when the parent record is accessible
-     * through the recordWhereClause
-     *
-     * @param int $recordUid
-     * @param string $parentWhereClause
-     * @return array
-     */
-    protected function getPageOverlayRecordIfParentIsAccessible($recordUid, $parentWhereClause)
-    {
-        $overlayRecord = (array)BackendUtility::getRecord('pages_language_overlay', $recordUid, '*');
-        $pageRecord = (array)BackendUtility::getRecord('pages', $overlayRecord['pid'], '*', $parentWhereClause);
-
-        if (empty($pageRecord)) {
-            return [];
-        }
-
-        return $overlayRecord;
-    }
 }
